test: move tests/lib/anta.py
