+++ tests/units/anta_tests/routing/test_bgp.py
@@ -25,7 +25,7 @@ from anta.tests.routing.bgp import (
     VerifyBGPTimers,
     VerifyEVPNType2Route,
 )
-from tests.lib.anta import test  # noqa: F401; pylint: disable=W0611
+from tests.units.anta_tests import test  # noqa: F401; pylint: disable=W0611
 
 DATA: list[dict[str, Any]] = [
     {
