+++ tests/units/anta_tests/routing/test_isis.py
@@ -20,7 +20,7 @@ from anta.tests.routing.isis import (
     VerifyISISSegmentRoutingTunnels,
     _get_interface_data,
 )
-from tests.lib.anta import test  # noqa: F401; pylint: disable=W0611
+from tests.units.anta_tests import test  # noqa: F401; pylint: disable=W0611
 
 DATA: list[dict[str, Any]] = [
     {
