+++ tests/units/anta_tests/test_aaa.py
@@ -16,7 +16,7 @@ from anta.tests.aaa import (
     VerifyTacacsServers,
     VerifyTacacsSourceIntf,
 )
-from tests.lib.anta import test  # noqa: F401; pylint: disable=unused-import
+from tests.units.anta_tests import test  # noqa: F401; pylint: disable=unused-import
 
 DATA: list[dict[str, Any]] = [
     {
