+++ tests/units/anta_tests/test_hardware.py
@@ -16,7 +16,7 @@ from anta.tests.hardware import (
     VerifyTransceiversManufacturers,
     VerifyTransceiversTemperature,
 )
-from tests.lib.anta import test  # noqa: F401; pylint: disable=W0611
+from tests.units.anta_tests import test  # noqa: F401; pylint: disable=W0611
 
 DATA: list[dict[str, Any]] = [
     {
