+++ tests/units/anta_tests/test_system.py
@@ -18,7 +18,7 @@ from anta.tests.system import (
     VerifyReloadCause,
     VerifyUptime,
 )
-from tests.lib.anta import test  # noqa: F401; pylint: disable=W0611
+from tests.units.anta_tests import test  # noqa: F401; pylint: disable=W0611
 
 DATA: list[dict[str, Any]] = [
     {
